Add more class diagrams
Class diagrams added for:
- Class
- Difficulty
- EncryptedPassword
- Worksheet

Eric, this Worksheet class is different from the ones you wrote in the
sense that it’s the object that is passed back and forth to the
database and used to re-create existing worksheets.

Ethan is working on the remaining classes: User, Teacher, Student,
Admin. I believe those are all the classes we’ll need. Let me know if
we missed anything!

Narrative and relationship diagram to come.