Fix duplicate relationship diagram and squeezed text in diagrams
I re-added some of the class diagrams to both the report and the
powerpoint and resized them in a way that keeps the text easy to read
and takes advantage of empty space.
Also, one the of the relationship diagrams was repeated, so I replaced
it with the correct one in both the report and the presentation.