Created classes based on class diagrams from phase III
Also updated existing classes with details from the class diagrams.
However, these may change soon. I’m considering that instead of loading
and storing a teacher’s list of worksheets as soon as they log in,
maybe only load them as we need them, for example when viewing their
history. There are more examples of this. We can discuss these
decisions on Sunday.