Add teacher registration panel and admin menu panel
Tomorrow I will be working on integrating these panels with the main
frame so the program runs faster and more smoothly. I would like to
avoid re-creating a panel if it has already been created so the GUI
isn’t too slow, so I will be writing the code that enforces all that.
I will also be working on Student panels and will help Olga with the
Teacher panels.