Integrate all Admin panels and advanced worksheets
Made slight updates to the internal structure. FWCConfigurator now
needs to keep track of what the current page the GUI is displaying is,
so that the help utility can later be context-sensitive. Admin panels
still need some tweaking to look good, but otherwise appear to be
functional, but we can’t know for sure until we hook up the database…