Finish database connection and other updates
Had to disable username editing as at this point there’s no easy way of
holding on to the old username to identify the user in the database to
update.
NOTE: database connection will only work for you if you set your local
MySQL password at the top of the FWCDatabaseConnection class.
I am testing, fixing bugs, and making the GUI look better.