Update to final report material
I discovered a bug right before submitting, but wasn’t able to fix it
in time. If you create a new class as a teacher and then register a new
student right afterwards, the new class will not be listed in the class
selection in the student registration page. I also don’t have it
restricting the teacher from creating a student if they haven’t created
a class first. All things I can easily fix before the presentation, but
wanted to let you know.